+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add app::event_mouse_button();
 - Add enums::Mouse.
 - Fix docs.
+- impl IntoIter for MenuItem.
 
 ## [0.14.3] - 2021-02-08
 ### Changes